Fix housing_units.baths type: f32 to match the `real` column
The `baths` column is created as `real` (single precision) by migration
000015, but the entity, CreateUnitParams, and the controller request all
declared it `Option<f64>`. On Postgres, SeaORM/sqlx binds an f64 as a float8
parameter, which Postgres rejects against a float4 (`real`) column — so
POST /api/housing/nodes/{id}/units returned 500 and the
`assign_and_vacate_requeues_unit` test failed. SQLite's dynamic typing hid
the mismatch locally, which is why it only showed in CI.

Align the Rust type with the actual column (`Option<f32>`). Bath counts
(1.0, 1.5, 2.5) are represented exactly in single precision, so there is no
loss. Verified the full suite passes against Postgres (32/32 request tests).

+++ src/controllers/housing.rs
@@ -49,7 +49,7 @@ pub struct VacateUnitRequest {
 pub struct CreateUnitRequest {
     pub unit_number: String,
     pub beds: Option<i32>,
-    pub baths: Option<f64>,
+    pub baths: Option<f32>,
     pub sqft: Option<i32>,
     pub notes: Option<String>,
 }
